Modified environment. Add server dependencies.
