+++ .gitignore
@@ -2,15 +2,16 @@
 
 # dependencies
 /node_modules
-/.pnp
+client/node_modules
+client/.pnp
 .idea
 .pnp.js
 
 # testing
-/coverage
+client/coverage
 
 # production
-/build
+client/build
 
 # misc
 .DS_Store
