+++ client/public/index.html
@@ -10,7 +10,7 @@
       content="Web site created using create-react-app"
     />
     <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap" />
-    <link rel="stylesheet" href="./node_modules/normalize.css/">
+    <link rel="stylesheet" href="../node_modules/normalize.css">
     <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
 
     <title>ЛООП Защита</title>
