+++ client/src/components/card/Card.module.scss
