+++ client/src/layout/header/components/hero/Hero.jsx
