+++ client/src/pages/dashboard/Dashboard.module.scss
@@ -1,0 +1,5 @@
+@import "../../scss/variables";
+
+.dash {
+  display: flex;
+}
