+++ client/src/pages/dashboard/components/MenuSection/MenuSection.module.scss
@@ -1,0 +1,9 @@
+@import "../../../../scss/variables";
+
+.menuSection {
+  min-width: calc(100% - #{$dashboardMenuWidth});
+  height: 100vh;
+  background: #f1f1f1;
+  padding: 10px;
+  font-size: 0.8rem;
+}
