--- src/pages/News.jsx
@@ -1,1 +1,0 @@
-'use strict'
